fix(memory): resolve CodeQL warnings, redundant instanceof checks, and optimize bypassWal recovery lifecycle
- Fix CodeQL narrowing cast warning in AbstractAppendMemory
- Optimize bypassWal lifecycle in WalRecoveryDispatcher to wrap recovery block instead of toggling per event
- Remove redundant instanceof AbstractTierStore checks in SpectorMemoryFactory
- Remove unused variables in NamespaceRegistryTest

+++ memory/spector-memory/src/main/java/com/spectrayan/spector/memory/kernel/shape/AbstractAppendMemory.java
@@ -71,7 +71,7 @@ public abstract class AbstractAppendMemory<L extends MemoryLayout>
         MemorySegment.copy(bytes, 0, segment(), writeOffset + 4, len);
 
         long payloadOffset = count + 4;
-        count += 4 + len;
+        count += (int) (4 + len);
         persistCount();
         return payloadOffset;
     }
